Skeleton, Navigation and Contact Form
Basically I've got a skeleton of a react site with working router, the navigation component works and loads the appropriate pages, the contact page has a working form that fulfills the assignment requirements. There is a very barebones About page that could work. Portfolio and Resume pages need to be done, along with Footer, Header and Project components. Styling is also lacking currently.